# docs/swagger.yaml
definitions:
  models.DeleteUserPayload:
    properties:
      id:
        type: integer
    type: object
  models.EditUserPayload:
    properties:
      id:
        type: integer
      name:
        type: string
      patronymic:
        type: string
      surname:
        type: string
    type: object
  models.SaveUserPayload:
    properties:
      name:
        type: string
      patronymic:
        type: string
      surname:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      description: Retrieve users with optional filters
      parameters:
      - description: Filter by name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by surname (partial match)
        in: query
        name: surname
        type: string
      - description: Filter by patronymic (partial match)
        in: query
        name: patronymic
        type: string
      - description: Minimum age
        in: query
        name: ageFrom
        type: integer
      - description: Maximum age
        in: query
        name: ageTo
        type: integer
      - description: Filter by sex (male/female)
        in: query
        name: sex
        type: string
      - description: Filter by country code (partial match)
        in: query
        name: country
        type: string
      - description: Pagination limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get filtered users
      tags:
      - users
  /add:
    post:
      consumes:
      - application/json
      description: Add new user with data enrichment
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SaveUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - users
  /delete:
    post:
      consumes:
      - application/json
      description: Delete user by ID
      parameters:
      - description: Delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a user
      tags:
      - users
  /edit:
    post:
      consumes:
      - application/json
      description: Update user information
      parameters:
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EditUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update a user
      tags:
      - users
swagger: "2.0"